% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basegraphics.R
\name{plot_done}
\alias{plot_done}
\title{Let rmote know that a base R plot is complete and ready to serve}
\usage{
plot_done()
}
\description{
Let rmote know that a base R plot is complete and ready to serve
}

