% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{plot_expression}
\alias{plot_expression}
\title{Wrap a expression that creates a base R plot as a plot_expression}
\usage{
plot_expression(x)
}
\arguments{
\item{x}{an expression}
}
\description{
Wrap a expression that creates a base R plot as a plot_expression
}
\examples{
\dontrun{
start_rmote(basegraphics = FALSE)
plot_expression(expression({
  plot(10:1)
  abline(0, 1)
}))
}
}
