% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help.R
\name{print.help_files_with_topic}
\alias{print.help_files_with_topic}
\title{Print help page to servr}
\usage{
print.help_files_with_topic(x, ...)
}
\arguments{
\item{x}{help object}

\item{\ldots}{additional parameters}
}
\description{
Print help page to servr
}

