% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidgets.R
\name{print.htmlwidget}
\alias{print.htmlwidget}
\title{Print an htmlwidget to servr}
\usage{
print.htmlwidget(x, ...)
}
\arguments{
\item{x}{htmlwidget object}

\item{\ldots}{additional parameters}
}
\description{
Print an htmlwidget to servr
}

