% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{print.plot_expression}
\alias{print.plot_expression}
\title{Print plot_expression plot to servr}
\usage{
\method{print}{plot_expression}(x, ...)
}
\arguments{
\item{x}{plot_expression object}

\item{\ldots}{additional parameters}
}
\description{
Print plot_expression plot to servr
}
