% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trellis.R
\name{print.trellis}
\alias{print.trellis}
\title{Print trellis plot to servr}
\usage{
print.trellis(x, ...)
}
\arguments{
\item{x}{trellis object}

\item{\ldots}{additional parameters}
}
\description{
Print trellis plot to servr
}

