% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trellis.R
\name{rmote_device}
\alias{rmote_device}
\title{Set device parameters for traditional plot output}
\usage{
rmote_device(type = c("png", "pdf"), filename = NULL, retina = TRUE, ...)
}
\arguments{
\item{type}{either "png" or "pdf"}

\item{filename}{optional name for file (should have no extension and no directories)}

\item{retina}{if TRUE and type is "png", the png file will be plotted at twice its size and shown at the original size, to look better on high resolution displays}

\item{\ldots}{parameters passed on to either \code{\link[grDevices]{png}} or \code{\link[grDevices]{pdf}} (such as width, height, etc.)}
}
\description{
Set device parameters for traditional plot output
}

