% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmote.R
\name{rmote_mode}
\alias{rmote_mode}
\title{Set the rmote mode to be on}
\usage{
rmote_mode(on = TRUE, server_dir = file.path(tempdir(), "rmote_server"),
  help = TRUE, traditional = TRUE, htmlwidgets = TRUE)
}
\arguments{
\item{on}{should rmote act as if the rmote server is on?}

\item{server_dir}{directory where rmote server is running}

\item{help}{(logical) send results of `?` to servr}

\item{traditional}{(logical) send traditional lattice / ggplot2 plots to servr}

\item{htmlwidgets}{(logical) send htmlwidgets to servr}
}
\description{
This is useful if there is another R session on the server with an rmote server running that the current session would like to make use of.
}
\note{
This is useful when rmote is serving from a
}

