% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_html.R
\name{serve_rmd_html}
\alias{serve_rmd_html}
\title{Serve rmarkdown html report as a webpage}
\usage{
serve_rmd_html(dir, file.name, ...)
}
\arguments{
\item{dir}{directory that contains html report(s) that should be served}

\item{file.name}{name of the html file (report) that should be served}

\item{\ldots}{additional parameters}
}
\description{
Serve rmarkdown html report as a webpage
}

