% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmote.R
\name{start_rmote}
\alias{start_rmote}
\title{Initialize a remote servr}
\usage{
start_rmote(server_dir = file.path(tempdir(), "rmote_server"), port = 4321,
  daemon = TRUE, help = TRUE, traditional = TRUE, htmlwidgets = TRUE)
}
\arguments{
\item{server_dir}{directory to launch servr from}

\item{port}{port to run servr on}

\item{daemon}{logical - should the server be started as a daemon?}

\item{help}{(logical) send results of `?` to servr}

\item{traditional}{(logical) send traditional lattice / ggplot2 plots to servr}

\item{htmlwidgets}{(logical) send htmlwidgets to servr}
}
\description{
Initialize a remote servr
}

