% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trellis.R
\name{print.ggplot}
\alias{print.ggplot}
\title{Print ggplot2 plot to servr}
\usage{
print.ggplot(x, ...)
}
\arguments{
\item{x}{ggplot object}

\item{\ldots}{additional parameters}
}
\description{
Print ggplot2 plot to servr
}

