% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{print.help_files_with_topic}
\alias{print.help_files_with_topic}
\title{Print help page to servr}
\usage{
\method{print}{help_files_with_topic}(x, ...)
}
\arguments{
\item{x}{help object}

\item{\ldots}{additional parameters}
}
\description{
Print help page to servr
}

