% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{print.trellis}
\alias{print.trellis}
\title{Print trellis plot to servr}
\usage{
\method{print}{trellis}(x, ...)
}
\arguments{
\item{x}{trellis object}

\item{\ldots}{additional parameters}
}
\description{
Print trellis plot to servr
}

