% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmote.R
\name{rmote_on}
\alias{rmote_on}
\title{Set the rmote mode to be on}
\usage{
rmote_on(server_dir, help = TRUE, graphics = TRUE, htmlwidgets = TRUE)
}
\arguments{
\item{server_dir}{directory where rmote server is running}

\item{help}{(logical) send results of `?` to servr}

\item{graphics}{(logical) send traditional lattice / ggplot2 plots to servr}

\item{htmlwidgets}{(logical) send htmlwidgets to servr}
}
\description{
Set the rmote mode to be on
}
\note{
This is useful when running multiple R sessions on a server, where all will serve the same rmote process.  It is not necessary to call this in the same session on which \code{\link{start_rmote}} has been called, but on any other R sessions.
}

