% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{rmote_server_init}
\alias{rmote_server_init}
\title{Initialize a remote servr}
\usage{
rmote_server_init(server_dir = file.path(tempdir(), "rmote_server"),
  port = 4321, use_live = FALSE, help = TRUE, raster = TRUE,
  htmlwidgets = TRUE)
}
\arguments{
\item{server_dir}{directory to launch servr from}

\item{port}{port to run servr on}

\item{use_live}{should live.js be used to live reload the page when something new is plotted?}

\item{help}{(logical) send results of `?` to servr}

\item{raster}{(logical) send lattice / ggplot2 plots to servr}

\item{htmlwidgets}{(logical) send htmlwidgets to servr}
}
\description{
Initialize a remote servr
}
\note{
live.js has some issues with some browsers so by default it is not enabled, meaning that you will have to manually refresh your page.
}

